requirements.txt
numpy<2
pandas
networkx
matplotlib
scikit-learn
SimpleITK
pyradiomics
nibabel